Migrate next-auth to Next.js 13
